Create a custom build of tiddlywebwiki with external javascript
in a perfect world we'd be able to update the split.recipe to
instead of including the FileAdaptor include the 3 TiddlyWeb plugins
required for saving.

This isn't a perfect world as a broken tiddlywiki is outputted but
this is a good start.

Note that this special version of tiddlywiki doesn't include
deprecated javascript